Rework DT validation
Calling the dt-validate.py script is too slow to be called on each DT file
because loading the schema every time is slow. Rework validation to be a
new top level target.

This mostly reverts the prior way, but leaving it for now in case we
come up with a way to do this per file.

Signed-off-by: Rob Herring <robh@kernel.org>